requirements.txt
tensorflow>=2.8.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
optuna>=2.10.0
joblib>=1.1.0
tqdm>=4.62.0
python-binance>=1.0.16
configparser>=5.2.0
pytest>=7.0.0

